# client/src/api/openapi.json
{
  "components": {
    "schemas": {
      "CityLocation": {
        "properties": {
          "formatted_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formatted Address"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "place_id": {
            "title": "Place Id",
            "type": "string"
          }
        },
        "required": [
          "place_id",
          "lat",
          "lng",
          "formatted_address"
        ],
        "title": "CityLocation",
        "type": "object"
      },
      "CityPrediction": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "place_id": {
            "title": "Place Id",
            "type": "string"
          }
        },
        "required": [
          "place_id",
          "description"
        ],
        "title": "CityPrediction",
        "type": "object"
      },
      "CreateSessionRequest": {
        "properties": {
          "initial_query": {
            "title": "Initial Query",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "enum": [
              "solo",
              "group"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "initial_query"
        ],
        "title": "CreateSessionRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LocationInput": {
        "properties": {
          "formatted_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formatted Address"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "LocationInput",
        "type": "object"
      },
      "RestaurantOut": {
        "description": "User-facing details of one candidate; internal fields (scores, ids of members) stay out.",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "cuisine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuisine"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "maps_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maps Uri"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "place_id": {
            "title": "Place Id",
            "type": "string"
          },
          "price_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Level"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "rating_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Count"
          }
        },
        "required": [
          "place_id",
          "name",
          "address",
          "cuisine",
          "rating",
          "rating_count",
          "price_level",
          "distance_m",
          "maps_uri"
        ],
        "title": "RestaurantOut",
        "type": "object"
      },
      "ResultsResponse": {
        "description": "All candidates, best first. The client shows the first 3, then the rest on request.",
        "properties": {
          "restaurants": {
            "items": {
              "$ref": "#/components/schemas/RestaurantOut"
            },
            "title": "Restaurants",
            "type": "array"
          }
        },
        "required": [
          "restaurants"
        ],
        "title": "ResultsResponse",
        "type": "object"
      },
      "SearchRequest": {
        "description": "The setup answers, sent in one call after the last setup page.",
        "properties": {
          "after": {
            "enum": [
              "breakfast",
              "lunch",
              "dinner",
              "coffee_dessert",
              "drinks"
            ],
            "title": "After",
            "type": "string"
          },
          "dietary": {
            "default": "none",
            "enum": [
              "none",
              "vegetarian",
              "vegan",
              "halal"
            ],
            "title": "Dietary",
            "type": "string"
          },
          "minutes": {
            "maximum": 60.0,
            "minimum": 5.0,
            "title": "Minutes",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "walk",
              "drive"
            ],
            "title": "Mode",
            "type": "string"
          },
          "price_levels": {
            "default": [],
            "items": {
              "maximum": 4.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "title": "Price Levels",
            "type": "array"
          }
        },
        "required": [
          "after",
          "mode",
          "minutes"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "mode": {
            "enum": [
              "solo",
              "group"
            ],
            "title": "Mode",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "share_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Url"
          }
        },
        "required": [
          "session_id",
          "mode",
          "share_url"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hangry API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/places/autocomplete": {
      "get": {
        "operationId": "autocomplete_places_autocomplete_get",
        "parameters": [
          {
            "in": "query",
            "name": "input",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Input",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CityPrediction"
                  },
                  "title": "Response Autocomplete Places Autocomplete Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Autocomplete",
        "tags": [
          "places"
        ]
      }
    },
    "/places/cities/{place_id}": {
      "get": {
        "operationId": "city_location_places_cities__place_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "title": "Place Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityLocation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "City Location",
        "tags": [
          "places"
        ]
      }
    },
    "/sessions": {
      "post": {
        "operationId": "create_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/{session_id}/results": {
      "get": {
        "operationId": "session_results_sessions__session_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Results",
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/{session_id}/search": {
      "post": {
        "operationId": "search_session_sessions__session_id__search_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Session",
        "tags": [
          "sessions"
        ]
      }
    }
  }
}